+++ tests/mocks/i18n.ts
@@ -1,12 +1,13 @@
-import { vi } from "vitest";
 import { ref } from "vue";
-import { localisation } from "@tests/utils";
+import { localisation } from "@tests/localisation";
 
-vi.mock('vue-i18n', () => ({
-    useI18n: () => ({
-        t: localisation,
-        te: () => true,
-        locale: ref('fr')
-    }),
-    createI18n: () => {}
-}));
+export function mockI18n() {
+    return {
+        useI18n: () => ({
+            t: localisation,
+            te: () => true,
+            locale: ref('fr')
+        }),
+        createI18n: () => {}
+    };
+}